+++ src/types/rx-error.ts
@@ -84,9 +84,8 @@ export interface RxErrorParameters {
     readonly collection?: any;
     readonly database?: any;
     readonly compoundIndexes?: string[] | string[][] | undefined;
-
-
-
+    readonly indexes?: string[] | undefined;
+    readonly indexPath?: string;
 }
 
 /**