+++ test/unit/rx-schema.test.ts
@@ -14,13 +14,13 @@ import {
     createRxSchema
 } from '../../';
 import {
-    RxSchema,
     getIndexes,
     normalize,
     hasCrypt,
     getFinalFields,
     getPreviousVersions
 } from '../../dist/lib/rx-schema';
+import { RxJsonSchema } from '../../src/types';
 
 config.parallel('rx-schema.test.js', () => {
     describe('static', () => {
@@ -34,6 +34,7 @@ config.parallel('rx-schema.test.js', () => {
                 const indexes = getIndexes(schemas.bigHuman);
                 assert.ok(indexes.length > 1);
                 assert.deepStrictEqual(indexes[0], ['passportId']);
+                assert.deepStrictEqual(indexes[1], ['dnaHash']);
             });
             it('get sub-index', () => {
                 const indexes = getIndexes(schemas.humanSubIndex);
@@ -49,6 +50,12 @@ config.parallel('rx-schema.test.js', () => {
                 assert.ok(Array.isArray(indexes));
                 assert.ok(Array.isArray(indexes[0]));
                 assert.deepStrictEqual(indexes[0], ['passportId', 'passportCountry']);
+            });
+            it('get index from array', () => {
+                const indexes = getIndexes(schemas.humanArrayIndex);
+                assert.ok(Array.isArray(indexes));
+                assert.ok(Array.isArray(indexes[0]));
+                assert.deepStrictEqual(indexes[0], ['jobs.[].name']);
             });
         });
         describe('.checkSchema()', () => {
@@ -92,6 +99,41 @@ config.parallel('rx-schema.test.js', () => {
                 });
             });
             describe('negative', () => {
+                it('break when index defined at object property level', () => {
+                    assert.throws(() => SchemaCheck.checkSchema({
+                        version: 0,
+                        type: 'object',
+                        properties: {
+                            id: {
+                                type: 'string',
+                                primary: true
+                            },
+                            name: {
+                                type: 'string',
+                                index: true
+                            },
+                            job: {
+                                type: 'object',
+                                properties: {
+                                    name: {
+                                        type: 'string',
+                                        index: true
+                                    },
+                                    manager: {
+                                        type: 'object',
+                                        properties: {
+                                            fullName: {
+                                                type: 'string',
+                                                index: true
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                        },
+                        required: ['job']
+                    }), Error);
+                });
                 it('break when index is no string', () => {
                     assert.throws(() => SchemaCheck.checkSchema(schemas.nostringIndex), Error);
                 });
@@ -683,7 +725,6 @@ config.parallel('rx-schema.test.js', () => {
                 it('get firstLevel', async () => {
                     const schema = createRxSchema(schemas.human);
                     const schemaObj = schema.getSchemaByObjectPath('passportId');
-                    assert.strictEqual(schemaObj.index, true);
                     assert.strictEqual(schemaObj.type, 'string');
                 });
                 it('get deeper', async () => {
@@ -731,7 +772,7 @@ config.parallel('rx-schema.test.js', () => {
     });
     describe('issues', () => {
         it('#590 Strange schema behavior with sub-sub-index', async () => {
-            const schema = {
+            const schema: RxJsonSchema = {
                 version: 0,
                 type: 'object',
                 properties: {
@@ -746,14 +787,14 @@ config.parallel('rx-schema.test.js', () => {
                                 type: 'object',
                                 properties: {
                                     time: {
-                                        type: 'number',
-                                        index: true
+                                        type: 'number'
                                     }
                                 }
                             }
                         },
                     },
-                }
+                },
+                indexes: ['fileInfo.watch.time']
             };
             const db = await RxDB.create({
                 name: util.randomCouchString(10),
@@ -780,7 +821,7 @@ config.parallel('rx-schema.test.js', () => {
             db.destroy();
         });
         it('#620 indexes should not be required', async () => {
-            const mySchema = {
+            const mySchema: RxJsonSchema = {
                 version: 0,
                 type: 'object',
                 properties: {
@@ -792,15 +833,15 @@ config.parallel('rx-schema.test.js', () => {
                         type: 'string'
                     },
                     lastName: {
-                        type: 'string',
-                        index: true
+                        type: 'string'
                     },
                     age: {
                         type: 'integer',
                         minimum: 0,
                         maximum: 150
                     }
-                }
+                },
+                indexes: ['lastName']
             };
             // create a database
             const db = await RxDB.create({
@@ -819,25 +860,23 @@ config.parallel('rx-schema.test.js', () => {
             db.destroy();
         });
         it('#697 Indexes do not work in objects named "properties"', async () => {
-            const mySchema = {
+            const mySchema: RxJsonSchema = {
                 version: 0,
-                id: 'post',
                 type: 'object',
                 properties: {
                     properties: {
                         type: 'object',
                         properties: {
                             name: {
-                                type: 'string',
-                                index: true,
+                                type: 'string'
                             },
                             content: {
-                                type: 'string',
-                                index: true,
+                                type: 'string'
                             }
                         }
                     },
                 },
+                indexes: ['properties.name', 'properties.content']
             };
 
             // create a database
@@ -860,25 +899,23 @@ config.parallel('rx-schema.test.js', () => {
             db.destroy();
         });
         it('#697(2) should also work deep nested', async () => {
-            const mySchema = {
+            const mySchema: RxJsonSchema = {
                 version: 0,
-                id: 'post',
                 type: 'object',
                 properties: {
                     properties: {
                         type: 'object',
                         properties: {
                             name: {
-                                type: 'string',
-                                index: true,
+                                type: 'string'
                             },
                             properties: {
-                                type: 'string',
-                                index: true,
+                                type: 'string'
                             }
                         }
                     },
                 },
+                indexes: ['properties.name', 'properties.properties']
             };
 
             // create a database